*.py: foo is/== False -> not foo; foo is/== True -> foo
When something is already the essence of truth or falseness in Python,
you do not need to check it with comparison operators or the "is" operator.
Briefly, study:

    PEP 8 -- Style Guide for Python Code
    https://www.python.org/dev/peps/pep-0008/
    (search for "is True")

For a longer explanation,
read pages 300 through 305 of "Learning Python" 5th edition by Mark Lutz.

    Comparisons, Equality, and Truth
    The Meaning of True and False in Python

"foo is False" and "foo is True" are risky to use in Python.
The "is" operator in Python is not for comparing the values of objects.
The "is" operator indicates whether or not the objects are the same object.
Study the following:

    >>> a = list(range(10000000))
    >>> b = 12345678
    >>> c = b//2
    >>> d = c+c
    >>> b == d
    True
    >>> b is d
    False
    >>> id(b)
    138650856
    >>> id(d)
    259984896
    >>>

Notice that b and d have the same value, but are different objects.
It gets trickier.

    >>> b = 2
    >>> c = b//2
    >>> d = c+c
    >>> b == d
    True
    >>> b is d
    True
    >>> id(b)
    138487972
    >>> id(d)
    138487972
    >>>

b and d are the same object because of caching magic within the
main implementation of Python. This behavior could change.
So "foo is False" and "foo is True" are relying of the caching
behavior of Python, which could change.
Imagine what a nasty bug that would be to figure out.

Just for fun, study the following:

    >>> 0 is False
    False
    >>> 0 == False
    True
    >>> False is False
    True
    >>>

"foo == True" should be avoided because there can be multiple "true" values
other than True or 1.

"foo == False" is not an error for numeric values,
but is still eschewed. Use "not foo" instead.