*.py: Reduce SQL injection attack vulnerability. (Little Bobby Tables)
Study "Exploits of a Mom":

    https://www.xkcd.com/327/

The ad-hoc quoting for values embedded in SQL commands is risky.
If one has to do that, one should have quoting function(s)
to take care of that, but even those are prone to error also.

It is best to parameterize values passed to SQL!!!

Such is explained in:

    11.13. sqlite3 — DB-API 2.0 interface for SQLite databases
    https://docs.python.org/2/library/sqlite3.html