src/smart_module/rtc_interface.py: Rewrote some method docstrings.
The impetus for this was that docstrings using string
interpolation were ignored. They looked something like:

    '''... %s-byte ...''' % FOO_LEN

So I got rid of the string interpolation for docstrings.

Having hard coded values in the docstrings was unacceptable.
DRY is good.
Having to search for and change numbers is two places is like
fingers on a chalkboard, so I did not put them back.

I dithered much over whether to mention the names of the lengths,
or not.

    Earlier, I cited the various *_LEN names.
    Unfortunately, one does not get to see the values in the
    docstrings, but at least one knows what to look for.

    I chose to leave out most mention of lengths.
    That seems more Pythonic, but I could change my mind.

I tried to follow PEP 257 -- Docstring Conventions.
This is new to me, so scrutiny by more eyeballs will help.