src/smart_module/smart_module.py: Scheduler.prepare_jobs(): Speculative refactoring plus more.
Other eyeballs are needed to scrutinize this code.

I went far beyond mere refactoring, making many assumptions of which at least
one is likely to be wrong, so this code should be considered speculative.
If it works, that's fine.
If not, I hope it provokes thought about what we want .prepare_jobs() to do.

prepare_jobs() had repeated code patterns, so I refactored it,
then I tried to generalize it and minimize special cases.

Added capability to handle year stuff.
I do not know if that is valid to do or not.

Although schedule.every(job.interval).months.do(self.run_job, job) was
commented out, for consistency, that capability was added.
I do not know if that is valid to do or not.

What does a value of -1 for job.interval mean?

For job.interval <= -1,
does it make sense to do stuff for weeks, months, or years as I allow?