src/smart_module/smart_module.py: SmartModule.connect_influx: Fix bug (or create bug?).
This either fixes a bug or creates a bug.
If it fixes a bug, it should be merged.

It is about SmartModule.connect_influx() of src/smart_module/smart_module.py,
from jp-24 branch (52d42c9) excerpted as follows:

    233     def connect_influx(self, asset_context):
    ...
    236         databases = self.ifconn.get_list_database()
    237         for db in databases:
    238             if asset_context in db:
    239                 break
    240         else:
    241             self.ifconn.create_database(asset_context)
    242
    243         self.ifconn.switch_database(asset_context)
    244         return self.ifconn

Now to analyze it.

    236         databases = self.ifconn.get_list_database()

From http://influxdb-python.readthedocs.io/en/latest/api-documentation.html#influxdb.InfluxDBClient.get_list_database,
I understand that databases is a list of dictionaries
and that each dictionary has one item.
The item's key is u'name'.
The item's value is the database name.
The dox above give the example call and data:

>> dbs = client.get_list_database()
>> dbs
[{u'name': u'db1'}, {u'name': u'db2'}, {u'name': u'db3'}]

    237         for db in databases:

Hence each db is a dictionary.
The key is u'name'.
The value is the name of a database.

    238             if asset_context in db:

For dictionaries, the "in" operator lookds for keys, not values.
All the keys are u'name', so line 238 is equivalent to

if asset_context in u'name':

I get the feeling that it was desired for the test to
see if asset_context matched the name of a database.
I.e., to test if asset_context matched a value of the dictionary.
so I changed line 238 to:

    238             if asset_context in db.values():

So does this commit fix a bug or create a bug?