src/dumb_module/arduino/hapinode/hapinode.ino: loop(): Get rid of delay(100).
Woohoo! Got rid of the delay() in loop.

delay(100) is to be avoided.
- wastes power
- keeps CPU from doing useful stuff
- causes unnecessary latency for other things

This commit concludes the first batch of timekeeping changes.
Many changes have been made, so something has been broken.
It needs to be tested and debugged.
The current code keeps the CPU running all the time.

Obviously, a similar pattern of timekeeping code is repeated.
That begs for factoring into one place.
It would likely be done as a class if for no other reason
than to reduce globals and avoid naming collision worries.
That version will likely also keep the CPU running all the time.

Eventually, the CPU should be sleeping to save power when
there is nothing useful to do. I am very familiar with doing
this on raw microcontrollers, but do not know if the Arduino way of doing
things is compatible with that.

An example of code that sleeps the CPU when idle is
https://github.com/james-prior/ccc2016/tree/master/step08